Move utils to separate file
